Add option to switch lex_state per match
This could lead to unexpected bugs if `@next_token` was filled using one
parser and lookahead is called with an incompatible parser, so I've made
that an error.  All matches that set lex_state should be immediately
preceded by a "match" or "shift_token".

Probably better would be to push/pop on a `lex_state` stack.