Use ruby2_keywords for authenticator forwarding
Without this, authenticators need to accept keywords for compatibility
with older ruby versions.  Ruby 2.6 will send an empty hash for
`**properties`, and that can crash (or be interpreted as an optional
argument) if the authenticator doesn't accept keyword arguments.

Actually... is that a bad thing?  Maybe the error will motivate people
to upgrade to ruby 2.7+ or contribute their custom authenticators to
net-imap? 😉

It's easy to stay compatible with `ruby2_keywords`, so let's do that.