Fixes "bundle exec rake", clash with test/unit
When multiple versions of "test/unit" are installed, the Gemfile.lock
can be out of sync with "test/unit" version selected without bundler.
When that happens, "bundle exec rake" fails, even when "rake" passes.

This can generally be fixed by simply running "bundle update" so the
lockfile has the latest version, too.  And the lockfile is gitignored so
this doesn't affect CI.  But, the error can be confusing and waste time.