📚 Add "rake ghpages" for publishing rdoc
https://ruby.github.io/net-imap is now configured to automatically
publish the docs directory of the gh-pages branch.  This is the current
version of the rake task I used to create that branch.

This (or something like it) should be able to run from a Github action,
to automatically update after every new release is tagged.  But it can
be run manually until that is set up.

There are probably better ways to post the latest documentation to
https://ruby.github.io/net-imap, so this may be just a temporary
solution.