💥 Update responses methods for thread-safety
Calling `#responses` with no block is deprecated, as it can't be made
thread-safe and backward-compatible.  However, the deprecation warning
is commented out for now, until after I can check the test suites for
the rails and mail gems and submit PRs there if necessary.

For convenience, an optional `type` paramater has been added to
`#responses`, to yield only the array for that type of response.

Added `#clear_responses` as a convenience method for the common "read
and delete" pattern, and to provide a thread-safe responses
method that doesn't require a block.

The response handlers methods are now synchronized, but
`#response_handlers` now returns a frozen clone to alert users who had
manipulated it directly.  Hopefully very few are affected, but this
change is backwards incompatible.  N.b: there currently is no API for
inserting a response_handler into a particular position in the array.

Also, improve the yields_in_test_server_thread test helper: it no longer
reads `last_tag` from the block result, it sets a short `IO#timeout` in
ruby 3.2+, it uses a slightly longer timeout for overall test
completion, and it defines `sock.getcmd` so the tests can read more
intuitively than using a proc passed into the block.