✨ Add nameprep stringprep profile
This isn't used by or needed for Net::IMAP, but adding a new profile is
trivial with the new generic stringprep code.  Nameprep is used by IDNA,
and not useful without punycode.  But, with the addition of punycode,
Net::IMAP could also support IDNA hostnames.

On the other hand, IDNA and punycode should probably be added to `uri`,
not `net-imap`.  Perhaps `stringprep` should be extracted to its own
gem?