⚡️ Update flag parsing: FLAGS, LIST, PERMANENTFLAGS
Rather than treat all three identically, this uses a more strict parser
tailored to each one.  One of my earlier tests was invalid, so that has
also been fixed.

Rather than use `String#scan`, the flags lists are matched with a single
regexp, and the match is split then mapped.  In my benchmarks, this is
_slightly_ faster than using `String#scan`.