✨ Make text in resp-text optional (IMAP4rev2)
In RFC3501 (IMAP4rev1):

    resp-text       = ["[" resp-text-code "]" SP] text

In RFC9051 (IMAP4rev2):

    resp-text       = ["[" resp-text-code "]" SP] [text]

And in RFC9051 Appendix E:

     23.  resp-text ABNF non-terminal was updated to allow for empty text.

In the spirit of Appendix E. 23 (and based on some actual server
responses I've seen over the years), I've leniently re-interpreted this
as also allowing us to drop the trailing `SP` char after
`[resp-text-code parsable code data]`, like so:

    resp-text       = "[" resp-text-code "]" [SP [text]] / [text]

Actually, the original parser already _mostly_ behaved this way, because
the original regexps for `T_TEXT` used `*` and not `+`. But, as I
updated the parser in many other places to more closely match the RFCs,
that broke this behavior. This commit originally came _after_ many many
other changes.  While rebasing, I moved this commit first because that
simplified later commits.

Also:
* ♻️ Add `Patterns` module, to organize regexps.
* ♻️ Use `Patterns::CharClassSubtraction` refinement to simplify
  exceptions.
* ♻️ Add `ParserUtils::Generator#def_char_matchers` to define `SP`,
  `LBRA`, `RBRA`.
* ♻️ Add `ParserUtils#{match,accept}_re` to replace `TEXT`, `CTEXT` lex
  states.
* ♻️ Remove unused `lex_state` kwarg from match