✨ Add `#logout!` to combine logout and disconnect
A logout command that also disconnects, even if logout raises an
exception, and doesn't raise exceptions when already disconnected would
be very useful in situations where the connection MUST be gracefully
closed, e.g. for security or tests.

Although the change is relatively insignificant, changing #logout to
also ignore exceptions and immediately call #disconnect would be
backwards incompatible.  So a new method was created.