♻️ Extract `send_command_with_continuations` method
Abstracts running a command with a continuation handler.

In the base specifications (both RFC-3501 and RFC-9051), the only
command to need this is `AUTHENTICATE`.  The only other continuations
are for sending literals, which are simple command arguments.  Are there
any IMAP extensions that use continuations in this way?

Even with only one command, this still makes that code a little more
readable.  More importantly, it also simplifies the creation of a
generic SASL protocol adapter, for use by `net-sasl`, `net-pop`, and
others.