🔒 Update OAuthBearer to take one or two arguments
`OAuthBearerAuthenticator` was updated to allow two arguments, in order
to match the common `auth(username, secret)` parameter style.  In my
opinion, this makes the API a little bit confusing.  But the mechanism
only requires one argument, so it's natural to allow a single argument
version.  And this two argument version matches with how many clients
and applications seem to_assume_ the SASL mechanisms _always_ work.