♻️ Unify `response-data` with a single case stmt
Use `peek_re(RE_RESPONSE_TYPE)` and the `TAGGED_EXT_LABEL` regexp, to
unify numeric and non-numeric types in the same case stmt.  This is much
simpler than two tier version, but I don't have benchmarks as to whether
or not this version of the code performs any better.

Add aliases for nz-number.  Not validating that the numbers are within
their allowed range, for now at least.