⚡ Sort `response-data` case stmt
The case statement is sorted _roughly_ in the order of frequency...
but I need to collect more data to reliably make a claim about that.

In the meantime, this order made sense to me... more or less.

Current standard (both RFC3501 and RFC9051) responses are sorted first,
and deprecated, uncommon, unsupported, etc responses are sorted last.