🩹 Workaround buggy outlook.com address lists
Outlook.com sends an extraneous `SP` in between `address` values in its
envelope address lists: `env-{from,sender,reply-to,to,cc,bcc}`.  This
can't be parsed by the envelope performance tuning rewrite in #232, in
the v0.4.6 release.

I've had a private test for this quirk for a decade... but the test was
primarily testing a different (much stranger) BODYSTRUCTURE bug, so I
thought the extra SP was a transcription error and "fixed" the test...
Oops!

The example comes from a real response, but all addresses have been
replaced by the `faker` gem. :)