🔧🏷 Add type coercion for config attributes
This is implemented as another module included under the other two,
still based on overriding `attr_accessor`.  Types are only enforced by
the attr_writer methods.  Fortunately, rdoc isn't confused by keyword
arguments to `attr_accessor`, so the type can be added to that.

Currently, only `:boolean` and `Integer` are supported, but it should be
easy to add more.