🔧 Add Config#load_defaults
This is inspired by Rails::Application::Configuration#load_defaults in
rails.

`#load_defaults` resets the current config to behave like the versioned
default configuration for that version.  `#parent` will not be changed.

Some config attributes default to inheriting from their `#parent` (which
is usually Config.global) and are left unchanged, for example: `#debug`.