🔇 Fix ruby 2.7 keyword argument warning
There is _some_ value in testing that it works even without this change.
But the main purpose of the test is simply testing the conversion from a
hash parameter.  Forcing the kwargs to be empty is fine.

Other than backports, maybe, this is the last ruby 2.7 concession I'll
be making before we raise the minimum ruby requirement.  😄