💥🔥 Remove deprecated #client_thread attr_reader
I don't know what the original purpose for `#client_thread` was, and it
was deprecated by v0.4.0.  There's nothing particularly special about
the Thread that created the client instance.  Net::IMAP doesn't use it,
clients can use multiple threads, and users of Net::IMAP can track this
for themselves if they need to.