✨ Add FetchData attr accessors to UIDFetchData
This is accomplished by making UIDFetchData subclass FetchData.  I
instinctively prefer to extact a module or a shared superclass and put
all of the accessor methods on that.  I did it this way in order to
preserve the attribute accessor rdoc on FetchData.

But it feels wrong to me that UIDFetchData is subclassing FetchData:
* It's a small violation of Liskov's Substitution Principle, for #seqno
  (and, to a lesser extent, #uid).
* Any case statements or pattern matching that are looking for FetchData
  will also match UIDFetchData, which may not be correct.  It also
  complicates matching only UIDFetchData, because you need to ensure
  that when/in UIDFetchData comes before any when/in FetchData.

On the other hand, it's probably a good thing that FetchData#=== matches
UIDFetchData too.  Other than #seqno (which is prohibited by `UIDONLY`)
any code that works for FetchData today should work with UIDFetchData.