✨ QRESYNC: Add `vanished` kwarg to `#uid_fetch`
The `vanished` kwarg to `#uid_fetch` can be used to request a list all
of the message UIDs in `set` that have been expunged since
`changedsince`.  Setting `vanished` to true prepends a `VanishedData`
object to the returned array.  If the server does not return a
`VANISHED` response, an empty `VanishedData` object will still be added.

_The +QRESYNC+ capabability must be enabled._
[RFC7162](https://rfc-editor.org/rfc/rfc7162)

For example:
```ruby
  imap.enable("QRESYNC") # must enable before selecting the mailbox
  imap.select("INBOX")
  # the first value in the returned array is a VanishedData object
  vanished, *fetched = imap.uid_fetch(301..500, %w[FLAGS],
                                      changedsince: 12345,
                                      vanished: true)
```