🥅 Improve SequenceSet frozen errors
Previously, SequenceSet modification methods would raise a frozen error
when it updated its internal data structure (currently implemented as an
array of arrays), which was "can't modify frozen Array: [int, int]".
This was unexpected, confusing, and misleading.

This changes the SequenceSet modification methods to raise the expected
FrozenError _prior_ to attempting to modify the internal data structure.