✨ Responses with no args can return frozen dup hash
This adds a new `:frozen_dup` option to `config.responses_without_block`
which allows it to return a frozen copy of the responses hash, with each
response type array also being a frozen copy.  This seems like a
reasonable API so, now that we have a config option to control it, it
will tentatively become the new default in v0.6.