✅ Fix GH action for rubygems Trusted Publishing
I'm not sure why dependabot didn't suggest upgrading these.  But we need
to consider these actions as trustworthy, so I think that any security
risk due to using a version branch (rather than a checksum) is offset by
the hassle (and security risk!) of not automatically getting updates.

In particular, even if dependabot _did_ make PRs to upgrade these, it
wouldn't have made a PR for the `v0.4-stable` branch.