✨ Add Data polyfill for ruby 3.1
For new data structs, I'd prefer frozen by default and I don't want to
support the entire Struct API.  Data is perfect, but it's not available
until ruby 3.2.

So this adds a DataLite class that closely matches ruby 3.2's Data
class, and it can be a drop-in replacement for Data.  Net::IMAP::Data
is an alias for Net::IMAP::DataLite, so when we remove our
implementation, the constant will resolve to ruby's ::Data.

Ideally, we wouldn't define this on newer ruby versions at all, but that
breaks the YAML serialization for our test fixtures.

The tests (and some of the implementation) have been copied from the
polyfill-data gem and updated so that they use "Data" as it is resolved
inside the "Net::IMAP" namespace.  Copyright notices have been added to
the appropriate files to satisfy the MIT license terms.