📦 Add release.yml
I noticed that rdoc and irb have a release.yml.  This is slightly
different from theirs.  It has all of the categories that I've been
using for recent releases.  It does _not_ add emojis to the category
titles because that seems redundant when each individual PR has its own
emoji. 😉