🐛💥 Remove accidental `Data#attributes` method
`#attributes` isn't part of the core `Data` API, and it never should've
been added as a public method to our polyfill.

This is a breaking API change, _but only for users on ruby 3.1._

For users on ruby 3.2+, this is _not_ a breaking change.  For newer ruby
versions, `Data#attributes` wasn't defined and this fixes YAML encoding.

Since this only affects ruby 3.1 and this was only recently added (in
v0.5.2), it's unlikely anyone has written code that uses this method.
If they have, they'll need to convert it to `#to_h`.