🔧 ResponseParser config is mutable and non-global
When ResponseParser was initialized without any explicit config, it used
Config.global directly.  This meant that changes to `parser.config` were
also changes to the _global_ config!

When ResponseParser is initialized with a config hash, that creates
a frozen config object.  This meant that changes to `parser.config` were
impossible.

So, when the given config is global or frozen, we create a new config
and inherit from the given config.  We want to continue using the given
config as-is in other cases, so the client and its parser can share the
same exact config.