♻️ Drop ruby 3.1 Data polyfill
Depends on #538, which drops support for ruby 3.1. This code would only
run in ruby 3.1, so removing it is not be a breaking change.

Please note: TruffleRuby and JRuby support for Data is incomplete, but
the code doesn't currently distinguish between them and CRuby, so
they're the same level of broken as before. We're keeping the Data
tests, to help track full support in other ruby implementations.

For now, we're keeping Net::IMAP::Data only to support YAML encoding.
But the rest of the ::Data polyfill has been dropped. psych v5.2.5
handles Data natively, so DataLite will be dropped completely (in a
separate PR).