✨ Add configurable socket_read_limit
Sets the limit for how many bytes will be read from the socket at a
time.

When both socket_read_limit and max_response_size are non-nil, reads are
limited to the smaller of socket_read_limit and the remaining bytes for
that response.  When reading an IMAP `literal`, reads are also limited
to the remaining bytes in that `literal`

Please note that this only affects Net::IMAP's reads from the socket.
The underlying IO or Socket buffers and system calls will not be
directly affected by this limit.

This is complicated by the possibility that `gets(CRLF, limit)` could
stop in between a CR and a LF.