🐛 Fix inconsistently frozen SequenceSet#[] result
This maybe isn't actually documented very well (or at all...) but most
SequenceSet transform methods return a frozen result when +self+ is
frozen and a mutable result when +self+ is mutable.  Except +limit+
which always returns a frozen result.  And (before this commit) +slice+,
which inconsistently returned with matching frozen status when the
result wasn't empty, but always returned a frozen set when the result
_was_ empty.

Adding these tests exposed a much more significant bug:
`SequenceSet#xor` mutates the reciever.