♻️ Generate same stringprep tables with ruby 3.4
Ruby 3.4 changed its `Hash#inspect` representation (for the better).
So, every time I run rake (after switching to a branch that updates
mtimes on related files), this file is converted to the ruby 3.4+
format.

This commit normalizes the output.  It uses the older format simply to
keep the diff small.  Next time we update the stringprep code, we can
switch to the ruby 3.4+ format.