♻️ Short-circuit logout state transition
Not doing this with the other states, because the intention is to
eventually store small bits of information about each state on the
state objects, e.g: which mailbox was selected, what caused the logout.

For `logout`, the first attempt wins and shouldn't be overwritten.
For `selected`, the last attempt _should_ win.  (That said, commands
that affect connection state should be serialized.)