🧵 Short-circuit logout state transition
Eventually the states will store small bits of information about each
state on the state objects, e.g: which mailbox was selected or what
caused the logout.  So for `logout`, the first attempt wins and
shouldn't be overwritten.  For other states, like `selected`, the last
attempt _should_ win.

On the other hand, state transitions should be reworked so that:
* state transitioning commands cannot run simultaneously
* all state transitions are effected _inside_ the receiver thread