⚡ Speed up #normalized using #normalized? [🚧 TODO: move to after normal @string == nil]
Short-circuiting with this may be faster, because `#normalized?` doesn't
need to generate an array of entries in order to check the string.

On the other hand, when `@string` is not normal, this requires parsing
the string and then generating a new one (rather than just generating
one up-front).  It'll be a bigger speed-up when we stop storing
normalized strings.