✨ Abridged SequenceSet#inspect output
When using SequenceSet to store mailbox state for mailboxes with many
millions of messages, the last thing you want to do is accidentally
generate a string and output that!

So this establishes a maximum number of entries to output as an inspect
string, currently 512.  Above that number, we output only a much smaller
number of entries from the head and tail, currently 16 from each end.
When the sequence set has an internal denormalized string, we use
regexps to grab the first and last groups of entries from that.

A note on the implementation:  At first, I used a simple Regexp match
anchored to `\z` to grab the tail entries from the string.  But that was
shockingly slow even when the string isn't extremely large.

Next I benchmarked replacing `Regexp#match` with `String#rpartition` and
that performed fine, using the same naive regexps.  `String#rindex` was
also fine.  They are both slower than `Regexp#match` anchored to `\A`,
but not too bad and not dramatically slower based on string size.

Then I benchmarked replacing `\d+` with `\d{0,10}` or `[1-9]\d{1,9}`,
and those both made the regexp fast... but `Regexp.linear_time?` doesn't
like nested quantifiers.  Fully expanding both the `\d+` and the
`(entry){100}` keeps it fast _and_ satisfies `Regexp.linear_time?`.

But that regexp is UGLY! 🙁