⚡️ Don't store SequenceSet#string when normalized
By not duplicating the data in `@tuples` and `string` this change leads
to substantial memory savings for large sequence sets.

By quickly scanning the entries after a string is parsed, we can bypass
the merge algorithm for normalized strings, with a small penalty for
non-normalized strings.

By not caching `@string`, we save a little bit of wasted time when it
isn't used, and spend a little extra time when it's used more than once.