♻️ Compile Config.default using AttrVersionDefaults
Similar to the versioned defaults, this could be a source of conflicts
when merging, rebasing, or/cherry-picking feature and backport branches.
Now that all of the attributes specify their defaults, the current
version's default config can be compiled from that.

The sanity-check at the end of `config.rb` has been removed.  It was
there to protect against bad merges.  Now that `Config.default` is also
compiled from the version defaults added to each attr_accessor, the
tests in `test/net/imap/test_config.rb` are good enough.