✨ Add `SequenceSet#normalized?`
Returns whether `#string` is fully normalized: entries have been sorted,
deduplicated, and coalesced, and all entries are in normal form.  In
other words, `#normalized?` returns `true` if (and only if) `#string` is
equal to `#normalized_string`.  (`#entries` and `#elements` can be
identical for non-normalized strings.)

Because this is a new method, the benchmarks monkey-patch a naive
implementation of the method for prior versions of net-imap without it.