♻️ Rename SequenceSet's private 'minmaxes' attr
'#minmaxes` is currently an alias for `#set_data`, but an upcoming PR
will change how the `@set_data` is structured.  This prepares for that
by documenting which code expects an array (or enum) of `(min, max)`
tuples.  Some of this code could remain unchanged if `#minmaxes` were
changed to return an enumerator.