♻️ Rename SequenceSet's private 'runs' attr
This is an alias for `#set_data`, and is used to document code that
expects an array of "runs" but can be abstract about how those runs
are implemented.  An upcoming PR will replace `[min, max]` arrays with
`u32x2` packed integers, which dramatically reduces memory allocations
and speeds up _some_ operations (although it does slow down others).