🐛 Fix SequenceSet#delete?(num..num) to return set
According to the rdoc, using a range with `#delete?` should return a
sequence set (or nil), but the check came after importing the input, and
was based on whether the imported run contained multiple values.  So
deleting a single number range (e.g: `1..1` or `"1:1"`) would return a
number not a set.