🐛 Fix SequenceSet#max(n), cardinality < n <= size
This backports #580 to `v0.5-stable`.

When `SequenceSet#max(n)` is called with `n > cardinality`, it _should_
return a duplicate of the whole set.  But, `#max(n)` is implemented
using `#slice(-n..)`, and (copying the behavior of `Array`), when a
slicing starts from an out-of-range index, it returns `nil`.

It was using `-[count, size].min` to keep the index from going
out-of-range.  Prior to 0.6.0, `#size` is the same as `#count`, so it
gives incorrect results when the set contains an endless range.
`SequenceSet#cardinality` has not been backported to 0.5, so this makes
that calculation inline.

This change should also give a small performance boost, because it
bypasses the complexity of `#slice(range)` and just calls `#dup` (or
returns `self` when the set is frozen).