🐛 Fix `#responses()` freezing internal arrays
Now that `:frozen_dup` is the default behavior for `#responses` when
it's called without any arguments, a critical bug has become apparent:
it was not freezing the internal responses arrays directly, rather than
copies of them.  Freezing these arrays will, of course, lead to further
issues.

Ideally, code should be updated to use one of the other forms of
`#responses`, since this form is less efficient and also (intentionally)
incompatibile with old code that expects it to return mutable arrays.
But this is still a major bug.

Fixes #581, reported by @yurikoval.